Move errors types to their own folder.
