+++ src/errors/AuthorizationError.ts
