+++ src/errors/InvalidPasswordError.ts
